% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Download a dataset}
\usage{
get_dataset(id, save_local = FALSE)
}
\arguments{
\item{id}{The dataset id}

\item{save_local}{Either FALSE to directly download the dataset into R or a path to a directory to save a local copy}
}
\value{
A list containing the various parts of the dataset
}
\description{
Download a dataset
}
